+++ EngineSIU/EngineSIU.vcxproj
@@ -167,11 +167,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\BackgroundShaderCompileData.cpp" />
-    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\BackgroundShaderCompile.cpp" />
-    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderCompileUtility.cpp" />
-    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderFileMetadata.cpp" />
-    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderMetadataPtr.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompileData.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompile.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompileUtility.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderFileMetadata.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderMetadataPtr.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AmbientLightActor.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Editor\LevelEditor\SlateAppMessageHandler.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Editor\LevelEditor\SLevelEditor.cpp" />
@@ -297,11 +297,11 @@
     <ClCompile Include="EngineSIU\LightGridGenerator.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\BackgroundShaderCompileData.h" />
-    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\BackgroundShaderCompile.h" />
-    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderCompileUtility.h" />
-    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderFileMetadata.h" />
-    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderHotReload\ShaderMetadataPtr.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompileData.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompile.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompileUtility.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderFileMetadata.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderMetadataPtr.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Core\Container\Queue.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Core\Container\TSafeQueue.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AmbientLightActor.h" />
