+++ EngineSIU/EngineSIU/Engine/Source/Developer/ShaderCompileUtils/ShaderFileMetadata.h
