+++ EngineSIU/EngineSIU/Engine/Source/Developer/ShaderCompileUtils/ShaderMetadataPtr.cpp
